[CHANGE] Calculate proper juHash for jumpurl
This reverts revision 2930.

Revision 2930 removed the TYPO3 standard methods used to build a redirect url trying to
fix an uncaught TYPO3 Exception on jumpurls caused by the jumpurl security fix in TYPO3 4.5.24.
This commit restores the original code and adds a line to calculate the proper juHash for the jumpurl.

Resolves: #48971